test_requirements.txt
-r requirements.txt
mock
nose
spec
PyHamcrest
